API Reference

DECAES.T2mapOptions

DECAES.T2partOptions

DECAES.EPGdecaycurve

DECAES.T2mapSEcorr

DECAES.T2partSEcorr

DECAES.lcurve_corner

DECAES.lsqnonneg

DECAES.lsqnonneg_chi2

DECAES.lsqnonneg_gcv

DECAES.lsqnonneg_lcurve

DECAES.main

$T_2$-distribution mapping

DECAES.T2mapOptions — Type

T2mapOptions(; kwargs...)
T2mapOptions(image::Array{T,4}; kwargs...) where {T}

Options structure for T2mapSEcorr. This struct collects keyword arguments passed to T2mapSEcorr, performs checks on parameter types and values, and assigns default values to unspecified parameters.

Arguments

legacy::Bool

Perform T2-mapping using legacy algorithms. Default: false

MatrixSize::Tuple{Int64, Int64, Int64}

Size of first 3 dimensions of input 4D image. This argument has no default, but is inferred automatically as size(image)[1:3] when calling T2mapSEcorr(image; kwargs...).

nTE::Int64

Number of echoes in input signal. This argument is has no default, but is inferred automatically as size(image, 4) when calling T2mapSEcorr(image; kwargs...).

TE::Real

Interecho spacing (Units: seconds). This argument has no default.

nT2::Int64

Number of T2 times to estimate in the multi-exponential analysis. This argument has no default.

T2Range::Tuple{T, T} where T<:Real

Tuple of min and max T2 values (Units: seconds). This argument has no default.

T1::Real

Assumed value of T1 (Units: seconds). Default: 1.0

Threshold::Real

First echo intensity cutoff for empty voxels. Default: if !legacy 0.0 else 200.0 end

RefConAngle::Real

Refocusing pulse control angle for stimulated echo correction (Units: degrees). Default: 180.0

MinRefAngle::Real

Minimum refocusing angle for flip angle optimization (Units: degrees). Default: 50.0

nRefAngles::Int64

During flip angle optimization, goodness of fit is checked for up to nRefAngles angles in the range [MinRefAngle, 180]. The optimal angle is then determined through interpolation from these samples. Default: if !legacy 32 else 8 end

nRefAnglesMin::Int64

Initial number of angles to check during flip angle optimization before refinement near likely optima. Setting nRefAnglesMin equal to nRefAngles forces all angles to be checked. Default: if !legacy min(5, nRefAngles) else nRefAngles end

Reg::String

Regularization routine to use. One of "none", "chi2", "gcv", or "lcurve", representing no regularization, Chi2Factor-based Tikhonov regularization, Generalized Cross-Validation method, or L-Curve based regularization, respectively.

Chi2Factor::Union{Nothing, T} where T<:Real

Constraint on $\chi^2$ used for regularization when Reg == "chi2". Default: nothing

SetFlipAngle::Union{Nothing, T} where T<:Real

Instead of optimizing flip angle, use SetFlipAngle for all voxels (Units: degrees). Default: nothing

SaveResidualNorm::Bool

Boolean flag to include a 3D array of the $\ell^2$-norms of the residuals from the NNLS fits in the output maps dictionary. Default: false

SaveDecayCurve::Bool

Boolean flag to include a 4D array of the time domain decay curves resulting from the NNLS fits in the output maps dictionary. Default: false

SaveRegParam::Bool

Boolean flag to include 3D arrays of the regularization parameters $\mu$ and resulting $\chi^2$-factors in the output maps dictionary. Default: false

SaveNNLSBasis::Bool

Boolean flag to include a 5D (or 2D if SetFlipAngle is used) array of NNLS basis matrices in the output maps dictionary. Default: SetFlipAngle !== nothing

Silent::Bool

Suppress printing to the console. Default: false

Note

The 5D array that is saved when SaveNNLSBasis is set to true has dimensions MatrixSize x nTE x nT2, and therefore is typically extremely large. If the flip angle is fixed via SetFlipAngle, however, this is not an issue as only the unique nTE x nT2 2D basis matrix is saved.

See also:

T2mapSEcorr

DECAES.T2mapSEcorr — Function

T2mapSEcorr([io = stderr,] image::Array{T,4}; <keyword arguments>)
T2mapSEcorr([io = stderr,] image::Array{T,4}, opts::T2mapOptions{T})

Uses nonnegative least squares (NNLS) to compute T2 distributions in the presence of stimulated echos by optimizing the refocusing pulse flip angle. Records parameter maps and T2 distributions for further partitioning.

Arguments

image: 4D array with intensity data as (row, column, slice, echo)

A series of optional keyword argument settings which will be used to construct a T2mapOptions struct internally, or a T2mapOptions struct directly

Outputs

maps: dictionary containing parameter maps with the following fields:

Default Fields

"echotimes" Echo times of time signal (length nTE 1D array)

"t2times" T2 times corresponding to T2-distributions (length nT2 1D array)

"refangleset" Refocusing angles used during flip angle optimization (length nRefAngles 1D array by default; scalar if SetFlipAngle is used)

"decaybasisset" Decay basis sets corresponding to "refangleset" (nTE x nT2 x nRefAngles 3D array by default; nTE x nT2 2D array if SetFlipAngle is used)

"gdn": Map of general density = sum(T2distribution) (MatrixSize 3D array)

"ggm": Map of general geometric mean of T2-distribution (MatrixSize 3D array)

"gva": Map of general variance (MatrixSize 3D array)

"fnr": Map of fit to noise ratio = gdn / sqrt(sum(residuals.^2) / (nTE-1)) (MatrixSize 3D array)

"snr": Map of signal to noise ratio = maximum(signal) / std(residuals) (MatrixSize 3D array)

"alpha": Map of optimized refocusing pulse flip angle (MatrixSize 3D array)

Optional Fields

"resnorm": $\ell^2$-norm of NNLS fit residuals; see SaveResidualNorm option (MatrixSize 3D array)

"decaycurve": Signal decay curve resulting from NNLS fit; see SaveDecayCurve option (MatrixSize x nTE 4D array)

"mu": Regularization parameter used during from NNLS fit; see SaveRegParam option (MatrixSize 3D array)

"chi2factor": $\chi^2$ increase factor relative to unregularized NNLS fit; see SaveRegParam option (MatrixSize 3D array)

"decaybasis": Decay bases resulting from flip angle optimization; see SaveNNLSBasis option (MatrixSize x nTE x nT2 5D array, or nTE x nT2 2D array if SetFlipAngle is used)

distributions: T2-distribution array with data as (row, column, slice, T2 amplitude) (MatrixSize x nT2 4D array)

Examples

julia> image = DECAES.mock_image(MatrixSize = (100,100,1), nTE = 32); # mock image with size 100x100x1x32

julia> maps, dist = T2mapSEcorr(image; TE = 10e-3, nT2 = 40, T2Range = (10e-3, 2.0), Reg = "lcurve", Silent = true); # compute the T2-maps and T2-distribution

julia> maps
Dict{String, Any} with 10 entries:
  "echotimes"     => [0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09, 0.1  …  0.23, …
  "t2times"       => [0.01, 0.0114551, 0.013122, 0.0150315, 0.0172188, 0.0197244, 0.022594…
  "refangleset"   => [50.0, 54.1935, 58.3871, 62.5806, 66.7742, 70.9677, 75.1613, 79.3548,…
  "gdn"           => [185326.0 1.26487e5 … 1.85322e5 1.025e5; 1.48254e5 1.91871e5 … 1.5159…
  "fnr"           => [460.929 398.485 … 415.599 428.055; 316.05 340.206 … 372.7 410.487; ……
  "alpha"         => [165.348 164.657 … 163.031 166.735; 165.212 164.994 … 165.6 165.094; …
  "gva"           => [0.438218 0.330969 … 0.367929 0.350691; 0.287478 0.307593 … 0.3892 0.…
  "ggm"           => [0.0501681 0.0518976 … 0.0509003 0.0519268; 0.0535984 0.0528403 … 0.0…
  "snr"           => [357.763 317.439 … 322.451 340.88; 250.674 275.185 … 294.411 324.464;…
  "decaybasisset" => [0.0 0.0 … 0.0 0.0; 0.0 0.0 … 0.0 0.0; … ; 0.0 0.0 … 0.0 0.0; 0.0 0.0…

See also:

T2partSEcorr

lsqnonneg

lsqnonneg_chi2

lsqnonneg_gcv

lsqnonneg_lcurve

EPGdecaycurve

$T_2$-parts and the myelin water fraction

DECAES.T2partOptions — Type

T2partOptions(; kwargs...)
T2partOptions(t2dist::Array{T,4}; kwargs...) where {T}

Options structure for T2partSEcorr. This struct collects keyword arguments passed to T2partSEcorr, performs checks on parameter types and values, and assigns default values to unspecified parameters.

Arguments

legacy::Bool

Calculate T2-parts using legacy algorithms. Default: false

MatrixSize::Tuple{Int64, Int64, Int64}

Size of first 3 dimensions of input 4D T2 distribution. This argument is has no default, but is inferred automatically as size(t2dist)[1:3] when calling T2partSEcorr(t2dist; kwargs...).

nT2::Int64

Number of T2 times to use. This argument has no default.

T2Range::Tuple{T, T} where T<:Real

Tuple of min and max T2 values (Units: seconds). This argument has no default.

SPWin::Tuple{T, T} where T<:Real

Tuple of min and max T2 values of the short peak window (Units: seconds). This argument has no default.

MPWin::Tuple{T, T} where T<:Real

Tuple of min and max T2 values of the middle peak window (Units: seconds). This argument has no default.

Sigmoid::Union{Nothing, T} where T<:Real

Apply sigmoidal weighting to the upper limit of the short peak window in order to smooth the hard small peak window cutoff time. Sigmoid is the delta-T2 parameter, which is the distance in seconds on either side of the SPWin upper limit where the sigmoid curve reaches 10% and 90% (Units: seconds). Default: nothing

Silent::Bool

Suppress printing to the console. Default: false

See also:

T2partSEcorr

DECAES.T2partSEcorr — Function

T2partSEcorr([io = stderr,] T2distributions::Array{T,4}; <keyword arguments>)
T2partSEcorr([io = stderr,] T2distributions::Array{T,4}, opts::T2partOptions{T})

Analyzes T2 distributions produced by T2mapSEcorr to produce data maps of a series of parameters.

Arguments

T2distributions: 4D array with data as (row, column, slice, T2 amplitude)

A series of optional keyword argument settings which will be used to construct a T2partOptions struct internally, or a T2partOptions struct directly

Ouputs

maps: a dictionary containing the following 3D data maps as fields:

"sfr": small pool fraction, e.g. myelin water fraction (MatrixSize 3D array)

"sgm": small pool geometric mean T2 (MatrixSize 3D array)

"mfr": medium pool fraction, e.g. intra/extracellular water fraction (MatrixSize 3D array)

"mgm": medium pool geometric mean T2 (MatrixSize 3D array)

Examples

julia> dist = DECAES.mock_T2_dist(MatrixSize = (100,100,1), nT2 = 40); # mock distribution with size 100x100x1x40

julia> maps = T2partSEcorr(dist; T2Range = (10e-3, 2.0), SPWin = (10e-3, 25e-3), MPWin = (25e-3, 200e-3), Silent = true); # compute T2-parts maps

julia> maps # MWF is contained in maps["sfr"]
Dict{String, Any} with 4 entries:
  "sgm" => [0.0105638 0.0128673 … 0.0120685 0.0117591; 0.0108218 0.0139536 … 0.0118711 0.011…
  "mfr" => [0.879499 0.889248 … 0.886412 0.897167; 0.927041 0.887472 … 0.872513 0.89026; … ;…
  "sfr" => [0.120501 0.110752 … 0.113588 0.102833; 0.0729592 0.112528 … 0.127487 0.10974; … …
  "mgm" => [0.0621055 0.0617419 … 0.0612094 0.0615634; 0.0607907 0.0625588 … 0.0629157 0.062…

See also:

T2mapSEcorr

NNLS analysis

DECAES.lsqnonneg — Function

lsqnonneg(C::AbstractMatrix, d::AbstractVector)

Returns the nonnegative least-squares (NNLS) solution, X, of the equation:

\[X = \mathrm{argmin}_{x \ge 0} ||Cx - d||_2^2\]

Arguments

C::AbstractMatrix: Left hand side matrix acting on x

d::AbstractVector: Right hand side vector

Outputs

X::AbstractVector: NNLS solution

DECAES.lsqnonneg_chi2 — Function

lsqnonneg_chi2(C::AbstractMatrix, d::AbstractVector, Chi2Factor::Real)

Returns the regularized NNLS solution, X, that incurrs an increase in $\chi^2$ approximately by a factor of Chi2Factor. The regularized NNLS problem solved internally is:

\[X = \mathrm{argmin}_{x \ge 0} ||Cx - d||_2^2 + \mu^2 ||x||_2^2\]

where $\mu$ is determined by approximating a solution to the nonlinear equation

\[\frac{\chi^2(\mu)}{\chi^2_{min}} = \mathrm{Chi2Factor} \quad \text{where} \quad \chi^2_{min} = \chi^2(\mu = 0)\]

Arguments

C::AbstractMatrix: Decay basis matrix

d::AbstractVector: Decay curve data

Chi2Factor::Real: Desired $\chi^2$ increase due to regularization

Outputs

X::AbstractVector: Regularized NNLS solution

mu::Real: Resulting regularization parameter $\mu$

Chi2Factor::Real: Actual increase $\chi^2(\mu)/\chi^2_{min}$, which will be approximately equal to the input Chi2Factor

DECAES.lsqnonneg_gcv — Function

lsqnonneg_gcv(C::AbstractMatrix, d::AbstractVector)

Returns the regularized NNLS solution, X, of the equation

\[X = \mathrm{argmin}_{x \ge 0} ||Cx - d||_2^2 + \mu^2 ||L x||_2^2\]

where $L$ is the identity matrix and $\mu$ is chosen by the Generalized Cross-Validation (GCV) method.

Details of the GCV method and L-curve theory can be can be found in: Hansen, P.C., 1992. Analysis of Discrete Ill-Posed Problems by Means of the L-Curve. SIAM Review, 34(4), 561-580

Arguments

C::AbstractMatrix: Decay basis matrix

d::AbstractVector: Decay curve data

Outputs

X::AbstractVector: Regularized NNLS solution

mu::Real: Resulting regularization parameter $\mu$

Chi2Factor::Real: Resulting increase in $\chi^2$ relative to unregularized ($\mu = 0$) solution

DECAES.lsqnonneg_lcurve — Function

lsqnonneg_lcurve(C::AbstractMatrix, d::AbstractVector)

Returns the regularized NNLS solution, X, of the equation

\[X = \mathrm{argmin}_{x \ge 0} ||Cx - d||_2^2 + \mu^2 ||L x||_2^2\]

where $L$ is the identity matrix and $\mu$ is chosen by locating the corner of the "L-curve".

Details of L-curve theory and the Generalized Cross-Validation (GCV) method can be found in: Hansen, P.C., 1992. Analysis of Discrete Ill-Posed Problems by Means of the L-Curve. SIAM Review, 34(4), 561-580

Arguments

C::AbstractMatrix: Decay basis matrix

d::AbstractVector: Decay curve data

Outputs

X::AbstractVector: Regularized NNLS solution

mu::Real: Resulting regularization parameter $\mu$

Chi2Factor::Real: Resulting increase in $\chi^2$ relative to unregularized ($\mu = 0$) solution

DECAES.lcurve_corner — Function

lcurve_corner(f, xlow, xhigh)

Find the corner of the L-curve via curvature maximization using Algorithm 1 from:

A. Cultrera and L. Callegaro, “A simple algorithm to find the L-curve corner in the regularization of ill-posed inverse problems”. IOPSciNotes, vol. 1, no. 2, p. 025004, Aug. 2020, doi: 10.1088/2633-1357/abad0d

Extended phase graph algorithm

DECAES.EPGdecaycurve — Function

EPGdecaycurve(ETL::Int, flip_angle::Real, TE::Real, T2::Real, T1::Real, refcon::Real)

Computes the normalized echo decay curve for a MR spin echo sequence using the extended phase graph algorithm using the given input parameters.

Arguments

ETL::Int: echo train length, i.e. number of echos

flip_angle::Real: angle of refocusing pulses (Units: degrees)

TE::Real: inter-echo time (Units: seconds)

T2::Real: transverse relaxation time (Units: seconds)

T1::Real: longitudinal relaxation time (Units: seconds)

refcon::Real: value of Refocusing Pulse Control Angle (Units: degrees)

Outputs

decay_curve::AbstractVector: normalized echo decay curve with length ETL

Main entrypoint function

DECAES.main — Function

main(command_line_args = ARGS)

Entry point function for command line interface, parsing the command line arguments ARGS and subsequently calling one or both of T2mapSEcorr and T2partSEcorr with the parsed settings. See the Arguments section for available options.

See also:

T2mapSEcorr

T2partSEcorr